requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic==2.5.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
PyPDF2==3.0.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2